my-k8s-api/requirements.txt
anyio==4.9.0
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcrypt==3.2.2
blinker==1.7.0
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.8
cloud-init==24.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
cryptography==41.0.7
dbus-python==1.3.2
distro==1.9.0
distro-info==1.7+build1
dotenv==0.9.9
faiss-cpu==1.11.0
fastapi==0.95.0
filelock==3.18.0
fsspec==2025.3.2
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
googleapis-common-protos==1.70.0
groq==0.22.0
grpc-stubs==1.53.0.5
grpcio==1.71.0
h11==0.14.0
hf-xet==1.1.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.31.1
hyperlink==21.0.0
idna==3.10
incremental==22.10.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
kubernetes==26.1.0
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
netifaces==0.11.0
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
openai==1.75.0
packaging==25.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycurl==7.45.3
pydantic==1.10.2
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.2.3
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.5
pytz==2024.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
service-identity==24.1.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.26.1
sympy==1.14.0
systemd-python==235
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
Twisted==24.3.0
types-protobuf==4.25.0.20240417
types-requests==2.32.0.20250328
typing_extensions==4.13.2
ubuntu-pro-client==8001
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.27.1
uvloop==0.19.0
wadllib==1.3.6
websocket-client==1.8.0
wheel==0.42.0
wsproto==1.2.0
zope.interface==6.1
